### 2022-09-16 #4768 @danielskatz
👋 @mgm8 - please add a "Statement of Need" section to your paper, as discussed in https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.

You might also see if there is a DOI or URL that can be added for your first reference.

In addition, due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or so.
