+++ paper/paper.bib
@@ -49,7 +49,8 @@
     author       = {CCSDS},
     organization = {The Consultative Committee for Space Data Systems},
     address      = {CCSDS Secretariat, National Aeronautics and Space Administration, Washington, DC, USA},
-    year         = {2021}
+    year         = {2021},
+    url          = {https://public.ccsds.org/Pubs/132x0b3.pdf}
 }
 
 @misc{pyngham-doc,
