### 2022-11-03 #4768 @mgm8
@jbytecode I fixed the reference issue, now all entries of the bibtex file are cited/referenced along the manuscript.

About a potential reviewer, I have a suggestion: Yuri-M-Dias
