### 2022-11-06 #4768 @jbytecode
👋👋👋 Dear @gorlapraveen    👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**PyNGHam: A Python library of the NGHam protocol**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4768).

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a seperate thread. In that thread, there will be 23 check items for each single reviewer.

Thank you in advance!
