### 2022-11-08 #4768 @jbytecode
👋👋👋 Dear @0xCoto and @pritchardn  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**PyNGHam: A Python library of the NGHam protocol**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4768).

Thank you in advance!
