### 2022-11-08 #4768 @0xCoto
@jbytecode I took a quick look at the paper and the repository, and (while I'm not specifically familiar with the NGHam protocol itself) I would also be happy to contribute to the review of the work. 🙂
