### 2022-11-08 #4915 @editorialbot
[REVIEW]: PyNGHam: A Python library of the NGHam protocol
**Submitting author:** @mgm8 (<a href="http://orcid.org/0000-0003-4889-6021">Gabriel Marcelino</a>)
**Repository:** https://github.com/mgm8/pyngham
**Branch with paper.md** (empty if default branch): main
**Version:** v1.1.1
