### 2022-11-08 #4915 @0xCoto
@mgm8 - Could you please comment on how this library compares against conventional DSP tools (such as GNU Radio, whose versatility may permit similar processing pipelines)? Considering GNU Radio's increasing popularity in the field of satellite communications, I think it would be useful for readers to understand the cases where PyNGHam could replace/complement the functionality of GNU Radio-based modules, such as [`gr-nuts`](https://github.com/lofaldli/gr-nuts).

**Some initial minor comments** (language suggestions to improve the quality of the manuscript):
- Line 9: `improves considerably` → `considerably improves`
- Line 10: `of a communication link` → `of the communication link`
- Line 10: `uses` → `applications`
- Line 11: `in` → (remove)
- Line 11-12: `as the protocol of the radio links between stations on Earth and a satellite` → `supporting a reliable communication link between the ground station and the satellite`
- Line 12: `allows` → `enables`
- Line 15: `was developed` → `was originally developed` (to make its initial development clear and distinguish it from the work being presented in the paper)
- Line 16: `After that, it was` → `It was later on`
- Line 17: `in the` → `on the
- Line 17: `it's being used` → `is being used` or `has been used` (if not all missions are still active).
- Line 21-25: Consider linking relevant references for each satellite mission (where available)
- Line 26: `on the list above,` → `of the above list`
- Line 28: `which was the context` → (remove)
- Line 30-32: This paragraph needs rewording; "this way" does not quite seem to link the two sentences in a manner that appears fully relevant to one another.
- Line 33-35: `This library is already being used in the development of the satellites of the SpaceLab, specifically in the ground station software [@spacelab-decoder] [@spacelab-transmitter], that sends telecommand and receives data to/from the satellites.` → `This library is already being used in the design of satellites by SpaceLab, as part of the ground station software [@spacelab-decoder; @spacelab-transmitter] responsible for the uplink of telecommands and downlink of telemetry/payload data.`
- Line 37: `idea` → `intention`
- Line 38: `using an FEC` → `using a FEC`
- Line 38: `and a well-defined` → `on a well-defined`
- Line 40: `is used` → `is employed`
- Line 42: "GMSK" is an undefined acronym (should be "Gaussian minimum-shift keying")
- Line 42: `0xAAAAAAAA` → `` `0xAAAAAAAA` `` (i.e., surround sequence with "\`" to differentiate it from the font/format of the rest of the paragraph)
- Line 44-45: `The size tag field has seven different options, indicating seven different packet sizes, as described below.` → `The size tag field has seven different options, each corresponding to a unique packet size described in Table 1.`

_(More to come)_

Also, it might be useful to mention what "NGHam" stands for (i.e. "Next Generation Ham Radio") somewhere, as it does not appear to be a very well-known protocol. The reader currently has to head to the [Skagmo (2014)](https://github.com/skagmo/ngham) reference to figure this out.
