### 2022-11-10 #4915 @pritchardn
I'll be adding my comments as I go, I plan to have my review completed in the next two weeks :smile:

Documentation - Statement of Need: There is a space in the overview page for explaining the point of the protocol however the last sentence, where the statement of need actually is seems unfinished - ending with `, and so far.`

Consider elaborating a little more about why an easy-to-use Python-based implementation is useful (I think it is, but I'd like to see it written) - especially since this is a pure-python implementation, rather than a binding of the original implementation,
