### 2022-11-14 913f9e8
paper: General improvements according to JOSS reviewers
