+++ paper/paper.md
@@ -20,37 +20,36 @@ bibliography: paper.bib
 
 # Summary
 
-The PyNGHam library is a Python version of the original NGHam protocol library written in C by Jon Petter Skagmo (LA3JPA) [@ngham]. The NGHam protocol is an amateur radio protocol developed to be a modern version of the AX.25 protocol, with the main improvement being the addition of a forward error correction (FEC) algorithm, which improves considerably the robustness of a communication link. Nowadays, one of the main uses of this protocol is on small satellite projects (specifically in CubeSats), as the protocol of the radio links between stations on Earth and a satellite. This Python implementation allows easier integration and use of this protocol in computers and embedded devices.
-
+The PyNGHam library is a Python version of the original NGHam protocol library written in C by Jon Petter Skagmo (LA3JPA) [@ngham]. The NGHam protocol is an amateur radio protocol developed to be a modern version of the AX.25 protocol, with the main improvement being the addition of a forward error correction (FEC) algorithm, which considerably improves the robustness of the communication link. Nowadays, one of the main applications of this protocol is on small satellite projects (specifically CubeSats), supporting a reliable communication link between the ground station and the satellite. This Python implementation enables easier integration and use of this protocol in computers and embedded devices.
 
 # Statement of need
 
-The NGHam protocol was developed in the context of a CubeSat development at the Norwegian University of Science and Technology (NTNU) [@lofaldi2016]. After that, it was used by the SpaceLab team in the FloripaSat-1 CubeSat, and it's being used by all satellites of the group so far. A list of known satellites that used or plan to use the NGHam protocol is presented below:
+The NGHam protocol was originally developed in the context of a CubeSat development at the Norwegian University of Science and Technology (NTNU) [@lofaldi2016]. It was later on used by the SpaceLab team on the FloripaSat-1 CubeSat, and is being used by all satellites of the group so far. A list of known satellites that used or plan to use the NGHam protocol is presented below:
 
 * **FloripaSat-1** [@marcelino2020]
-* **GOLDS-UFSC (a.k.a. FloripaSat-2)**
-* **Catarina-A1**
-* **PION-BR1**
+* **GOLDS-UFSC (a.k.a. FloripaSat-2)** [@golds-ufsc]
+* **Catarina-A1** (Catarina Constellation) [@seman2022]
+* **PION-BR1** [@pion-br1]
 * **Aldebaran-1**
-* **NUTS-1**
+* **NUTS-1** [@lofaldi2016]
 
-The top three satellites on the list above, are satellites developed (or in development) by the same research group: the *Space Technology Research Laboratory* (SpaceLab), from *Universidade Federal de Santa Catarina* (Brazil), which was the context where this library was developed.
+The top three satellites of the above list, are satellites developed (or in development) by the same research group: the *Space Technology Research Laboratory* (SpaceLab), from *Universidade Federal de Santa Catarina* (Brazil), where this library was developed.
 
-From the knowledge of the author, there is no Python implementation of the protocol so far. This way, an implementation with a high-level language is useful for developing user applications to communicate with objects in orbit.
+From the knowledge of the author, there is no Python implementation of the protocol so far. An implementation with a high-level language facilitates the development of user applications to communicate with objects in orbit.
 
-This library is already being used in the development of the satellites of the SpaceLab, specifically in the ground station software [@spacelab-decoder] [@spacelab-transmitter], that sends telecommand and receives data to/from the satellites.
[…]
