+++ paper/paper.bib
@@ -73,3 +73,39 @@
     year        = {2022},
     url         = {https://github.com/spacelab-ufsc/spacelab-transmitter}
 }
+
+@misc{golds-ufsc,
+    author      = {{SpaceLab}},
+    title       = {{GOLDS-UFSC} Mission},
+    year        = {2022},
+    url         = {https://github.com/spacelab-ufsc/floripasat2-doc}
+}
+
+@article{seman2022,
+    author      = {Seman, Laio Oriel and Ribeiro, Brenda F. and Rigo, Cezar A. and Filho, Edemar Morsch and Camponogara, Eduardo and Leonardi, Rodrigo and Bezerra, Eduardo A.},
+    title       = {An Energy-Aware Task Scheduling for Quality-of-Service Assurance in Constellations of Nanosatellites},
+    journal     = {Sensors},
+    volume      = {22},
+    year        = {2022},
+    number      = {10},
+    issn        = {1424-8220},
+    doi         = {10.3390/s22103715}
+}
+
+@misc{pion-br1,
+    author      = {{PION Labs}},
+    title       = {Information for Communicating with {PION-BR1}},
+    year        = {2022},
+    url         = {https://github.com/pion-labs/PION-BR1/wiki/Radio-Information}
+}
+
+@inproceedings{aldebaran-1,
+  author    = {Carlos Alberto Rios Brito Junior, Edemar Morscsh Filho, Jose de Ribamar, Braga Pinheiro Junior and Luı́s Claudio de Oliveira Silva},
+  title     = {{The CubeSat mission Aldebaran I}},
+  booktitle = {},
+  series    = {},
+  year      = 2022,
+  pages     = {248--256},
+  publisher = {},
+  address   = {}
+}
