### 2022-11-15 #4915 @pritchardn
Documentation - Community Guidlines
I can see the CONTRIBUTING.md file, I would suggest:
* Mentioning its existence in the main README.md and main documentation (users just installing through pip may never look at the source).
* Mentioning how to reach out for support, not just bug reporting and code-contributions

Just makes things super clear and easily accessible :smile:
