### 2022-11-20 #4915 @0xCoto
@mgm8

> Do you think this explanation should be added to the paper and/or documentation?

Yes, this is definitely something worth addressing, as readers may be curious as to why they should give this package a try instead of GNU Radio. The potential applications and purpose of the package should be made clear in the manuscript.

You can also briefly state the advantages NGHam offers compared to other modulation/encoding schemes used in CubeSats, to emphasize the importance of having a Python-based solution for something like this.
