### 2022-11-20 3e4b63d
paper: Improvement in the statement of need according to JOSS reviewers
