+++ paper/paper.bib
@@ -109,3 +109,10 @@
   publisher = {},
   address   = {}
 }
+
+@misc{gr-nuts,
+    author      = {Løfaldli, André},
+    title       = {gr-nuts - {GNURadio} module for the {NUTS} project groundstation},
+    year        = {2022},
+    url         = {https://github.com/lofaldli/gr-nuts}
+}
