### 2022-11-20 4533d5c
doc: Adding contributing and contact information to the main page
