### 2022-11-20 cebe889
Improving the CONTRIBUTING file, and citing it in the README file
