### 2022-11-20 24c2b9f
doc: Fixing the contact information text
