### 2022-11-20 #4915 @mgm8
> @mgm8
>
> > Do you think this explanation should be added to the paper and/or documentation?
>
> Yes, this is definitely something worth addressing, as readers may be curious as to why they should give this package a try instead of GNU Radio. The potential applications and purpose of the package should be made clear in the manuscript.
>
> You can also briefly state the advantages NGHam offers compared to other modulation/encoding schemes used in CubeSats, to emphasize the importance of having a Python-based solution for something like this.

@0xCoto Ok, fixed in [3e4b63d](https://github.com/mgm8/pyngham/commit/3e4b63dcd9c75cbc86cc0f391a92cb1c95d1c2b6)!

> Documentation - Statement of Need: There is a space in the overview page for explaining the point of the protocol however the last sentence, where the statement of need actually is seems unfinished - ending with , and so far.
>
> Consider elaborating a little more about why an easy-to-use Python-based implementation is useful (I think it is, but I'd like to see it written) - especially since this is a pure-python implementation, rather than a binding of the original implementation,

@pritchardn I think this commit also solves your first suggestion.
