### 2022-11-20 #4915 @mgm8
> Documentation - Community Guidlines I can see the CONTRIBUTING.md file, I would suggest:
>
> * Mentioning its existence in the main README.md and main documentation (users just installing through pip may never look at the source).
> * Mentioning how to reach out for support, not just bug reporting and code-contributions
>
> Just makes things super clear and easily accessible smile

@pritchardn Thanks for your suggestions! I improved the references to the contributing instructions in [4533d5c](https://github.com/mgm8/pyngham/commit/4533d5c72be0c37702fd685413d85a1fc5efbcf1) and [cebe889](https://github.com/mgm8/pyngham/commit/cebe88915cfca5abec0f1e31f55eb060393d1f34).
