+++ _sources/index.rst.txt
@@ -9,6 +9,10 @@ PyNGHam Library Documentation
 
 The PyNGHam library is a Python version of the original NGHam protocol library written in C by Jon Petter Skagmo (LA3JPA).
 
+The project is fully open source and is available in a `GitHub repository <https://github.com/mgm8/pyngham>`_. All contributions are welcome! If you found a bug, developed a new feature, or want to improve the documentation, there are two ways to do so: open an issue describing the suggested modification, or by opening a pull request. More information are available in the `CONTRIBUTING file <https://github.com/mgm8/pyngham/blob/main/CONTRIBUTING.md>`_.
+
+Any questions or suggestions can also be addressed to Gabriel Mariano Marcelino <`gabriel.mm8@gmail.com <mailto:gabriel.mm8@gmail.com>`_>.
+
 Contents
 ========
 
