### 2022-11-25 #4915 @0xCoto
@mgm8 - Comments for the rest of the manuscript:

- "Statement of need" → "Statement of Need" (section capitalization is inconsistent)
- In Table 1, the _"up to"_ prefixes can be neglected, since the header _"**Max.** Data Size"_ already indicates the upper limit.
- Line 32-33: "GNURadio" → "GNU Radio" (and should be cited for those unfamiliar with the GNU Radio framework)
- Line 52:  "Inside the code block, there are two types of fields" → "The code block consists of two types of fields"
- Line 53-54: "The packet [...] divided ~~in~~ into four fields" needs to be reworded.
- Line 54: "Header" → "header"
- Line 54: Define the "CRC" acronym (Cyclic-Redundancy Code)
- Line 58-59: Since you're referencing CCSDS 131.0-B-3, could [CCSDS 131.0-B-3](https://public.ccsds.org/Pubs/131x0b3s.pdf) be more relevant than [CCSDS 132.0-B-3](https://public.ccsds.org/Pubs/132x0b3.pdf)? Also, could you please point me to where exactly the polynomial is brought up in the standard? In the manuscript, the polynomial is $x^8+x^7+x^5+x^3+1$. However, the only polynomial I see in the CCSDS 131.0-B-3 standard is on p. 29, which looks a bit different: $F(x) = x^8+x^7+x^2+x+1$. Are these two irrelevant?
- Line 61: "RS code block and gets the original content of the RS part of the packet" → "RS code block, ultimately retrieving the original content of the RS part of the packet."
- Line 62: "sequence" → "sequences"
- Line 62: "avoided," → "mitigated" (remove comma)
- Line 62-63: Change sentence to: "By scrambling the packets via bit transition, long sequences of ones or zeros are avoided."
- Line 63-64: This sentence can be shrunk down to just the citation alone, merged on the previous sentence (which does not have to be a separate paragraph, i.e. can merge with the previous sentence). Also, I do not see Section 8.3 in the cited standard (the last Section appears to be 6.6).
- Line 66: "Besides the RF transmissions using the NGHam protocol" → "While the NGHam protocol can be used to support wireless communications"
- Line 68: "there is no FEC algorithm involved, just" → "no FEC algorithm is involved, but merely"
- Line 68: "if" → "whether"
- Line 69: "or not" → remove
- Line 69: "Both RF and serial protocols" - _protocols_ does not sound like the appropriate term (since the protocol is NGHam). I would suggest "Both the RF and serial options"?
- Line 69: "following" → "subsequent"
- Line 75: What is meant by "repeating information"? Are you referring to a (communication) repeater/relay?
- Line 76: "the" → remove
- Line 78: "was inspired in part" → "was partly inspired"
- Line 80: "there is a class [...] protocol" → "a class [...] protocol is available"
- Line 80: "The" → "the"
- Line 82: Try to reword to avoid two nearby instances of ":" which introduces a bit of confusion. You could also shorten the beginning of line 83 to "packet encoding and decoding"
- Line 84: "or the payload data" unless this is different from "NGHam packet" you could wrap in parentheses or separate by a `/`, i.e., "NGHam packet/payload data"
- Line 86-87: Change to "The objective of this library is to offer an alternative to the original NGHam library, purely written in Python."
- Line 87: "This way, this" → "In this manner, such package"
- Line 87: "can be easily used in" → "can easily be used to carry out"
- Line 87-88: "packet decoding and encoding software, telecommunication classes, and so on" → "develop packet encoding and decoding tools, introduce telecommunication concepts to students, among others"
- Line 88-90: Sentence requires rewording.
- Line 90-91: This sentence belongs to the previous section _(Python Implementation)_. Also, "in its Sphinx" can be neglected (it is an unimportant detail for the reader). As for the reference, a direct link would be more useful.
- "Acknowledgement" → "Acknowledgements"
