### 2022-11-28 #4915 @pritchardn
Documentation: Automated Tests

I see that automated tests are present, however, they are not mentioned in the documentation. The `usage` section is appreciated, but please add an additional small section for running the supplied tests and explaining what functionality each test script is exercising. I don't mind if this is a top-level section or sub-section to `usage`.
