### 2022-11-29 #4915 @pritchardn
Functionality:

I've run the examples provided in the documentation with no issues, but I do have one comment.

The protocol description talks of the payload with respect to `bytes`, however, the encoding implementation only accepts lists of ints. I understand the list of ints represents the byte values, but a novice might be confused when passing a byte-array to the encode function of a protocol described to accept bytes to be greeted with an error.

Consider adding an example that converts an encoded string to a list of ints (easily done by passing the byte array to `list()`)  to make this step clear.
