### 2022-11-29 33013cf
paper: General improvements according to JOSS reviewers
