+++ paper/paper.bib
@@ -45,19 +45,12 @@
 }
 
 @manual{ccsds,
-    title        = {{TM} Space Data Link Protocol},
+    title        = {{TM} Synchronization and Channel Coding},
     author       = {CCSDS},
     organization = {The Consultative Committee for Space Data Systems},
     address      = {CCSDS Secretariat, National Aeronautics and Space Administration, Washington, DC, USA},
-    year         = {2021},
-    url          = {https://public.ccsds.org/Pubs/132x0b3.pdf}
-}
-
-@misc{pyngham-doc,
-    author      = {Marcelino, Gabriel Mariano},
-    title       = {{PyNGHam} Library Documentation},
-    year        = {2022},
-    url         = {https://mgm8.github.io/pyngham/}
+    year         = {2022},
+    url          = {https://public.ccsds.org/Pubs/131x0b4.pdf}
 }
 
 @misc{spacelab-decoder,
@@ -115,4 +108,10 @@
     title       = {gr-nuts - {GNURadio} module for the {NUTS} project groundstation},
     year        = {2022},
     url         = {https://github.com/lofaldli/gr-nuts}
+}
+
+@misc{gnu-radio,
+    author      = {GNU Radio Website},
+    year        = {2022},
+    url         = {https://www.gnuradio.org}
 }
