### 2022-11-29 25a60f9
Updating the name of some functions of the CRC library
