+++ pyngham/spp.py
@@ -23,7 +23,7 @@
 
 from enum import Enum
 import time
-from crc import CrcCalculator, Configuration
+from crc import Calculator, Configuration
 
 class SPPType(Enum):
     """
@@ -92,7 +92,7 @@ class PyNGHamSPP:
         # Payload
         pkt = pkt + pl
 
-        crc_val = CrcCalculator(Configuration(16, 0x1021, 0xFFFF, 0xFFFF, True, True)).calculate_checksum(pkt)
+        crc_val = Calculator(Configuration(16, 0x1021, 0xFFFF, 0xFFFF, True, True)).calculate_checksum(pkt)
         checksum = [(crc_val >> 8) & 0xFF, crc_val & 0xFF]
 
         # Start byte and CRC
@@ -213,7 +213,7 @@ below describes what is put into the payload of the general packet format.
 			# If received length has met target length (set in STATE_HEADER)
             if len(self._rx_buffer) == (2 + 1 + 1 + self._rx_buffer[3]):
                 # Check checksum
-                crc_val = CrcCalculator(Configuration(16, 0x1021, 0xFFFF, 0xFFFF, True, True)).calculate_checksum(self._rx_buffer[2:])
+                crc_val = Calculator(Configuration(16, 0x1021, 0xFFFF, 0xFFFF, True, True)).calculate_checksum(self._rx_buffer[2:])
 
                 self._state = SPPState.START
 
