### 2022-11-29 a0dfd94
Updating the name of some functions of the CRC library
