+++ pyngham/version.py
@@ -25,7 +25,7 @@ __author__      = "Gabriel Mariano Marcelino - PU5GMA"
 __copyright__   = "Copyright (C) 2022, Gabriel Mariano Marcelino"
 __credits__     = ["Gabriel Mariano Marcelino - PU5GMA"]
 __license__     = "LGPLv3"
-__version__     = "1.0.1"
+__version__     = "1.0.2"
 __maintainer__  = "Gabriel Mariano Marcelino - PU5GMA"
 __email__       = "gabriel.mm8@gmail.com"
 __status__      = "Production"
