### 2022-11-29 8105191
paper: Fixing the GNU Radio reference
