+++ paper/paper.bib
@@ -111,7 +111,7 @@
 }
 
 @misc{gnu-radio,
-    author      = {GNU Radio Website},
+    title       = {{GNU Radio} - The Free and Open Software Radio Ecosystem},
     year        = {2022},
     url         = {https://www.gnuradio.org}
 }
