### 2022-12-12 954ccb0
docs: Adding a description about the automated tests
