### 2022-12-12 e61787d
Converting data to byte array before calculating the CRC value
