### 2022-12-12 6cf1848
Adding the compatibility with bytes array and strings to the encode functions
