### 2022-12-12 67324bf
docs: Updating the usage instructions
