+++ _sources/packaging.rst.txt
@@ -1,17 +1,0 @@
-*********************
-Packaging the Project
-*********************
-
-This page presents the instructions to packaging the source files of the project.
-
-Generating a RPM package
-========================
-
-To generate an RPM package, execute the command below:
-
-::
-
-    python setup.py bdist_rpm
-
-
-If successful, the generated RPM package will be available in *dist/*.
