### 2022-12-12 #4915 @mgm8
> Documentation: Automated Tests
>
> I see that automated tests are present, however, they are not mentioned in the documentation. The `usage` section is appreciated, but please add an additional small section for running the supplied tests and explaining what functionality each test script is exercising. I don't mind if this is a top-level section or sub-section to `usage`.

@pritchardn I fixed this in [91323ac](https://github.com/mgm8/pyngham/commit/954ccb04d8aedb67b4ff51e7abbfcb85d91323ac).

> Functionality:
>
> I've run the examples provided in the documentation with no issues, but I do have one comment.
>
> The protocol description talks of the payload with respect to bytes, however, the encoding implementation only > accepts lists of ints. I understand the list of ints represents the byte values, but a novice might be confused when passing a byte-array to the encode function of a protocol described to accept bytes to be greeted with an error.
>
> Consider adding an example that converts an encoded string to a list of ints (easily done by passing the byte array to list()) to make this step clear.

Fixed in [cb7dd8f](https://github.com/mgm8/pyngham/commit/6cf184815005a37aca60d8ce0efc00ba1cb7dd8f) and [7d36b6f](https://github.com/mgm8/pyngham/commit/67324bf4751dab8a30d6755a7f48d19347d36b6f). Now the encode functions accepts byte arrays or strings as input.

Thanks for the suggestions.
