### 2022-12-14 mgm8/pyngham#20 @0xCoto
README example fails
It appears that the example on the README fails to run:

```py
Traceback (most recent call last):
  File "/home/apostolos/Desktop/joss_review/test_example.py", line 1, in <module>
    from pyngham import PyNGHam
  File "/home/apostolos/.local/lib/python3.10/site-packages/pyngham/__init__.py", line 24, in <module>
    from pyngham.pyngham import PyNGHam
  File "/home/apostolos/.local/lib/python3.10/site-packages/pyngham/pyngham.py", line 25, in <module>
    from crc import CrcCalculator, Configuration
ImportError: cannot import name 'CrcCalculator' from 'crc' (/home/apostolos/.local/lib/python3.10/site-packages/crc.py)
```
