### 2022-12-14 mgm8/pyngham#20 @jbytecode
Due to the JOSS submission: https://github.com/openjournals/joss-reviews/issues/4915
