### 2022-12-15 mgm8/pyngham#20 @mgm8
This is because of a recent update in the [crc library](https://pypi.org/project/crc/). In version 3.0 they changed the structure of the library.

I fixed this in [25a60f9](https://github.com/mgm8/pyngham/commit/25a60f9c7f9ab8aba860a91ad79d0ae03f1218b9) and [a0dfd94](https://github.com/mgm8/pyngham/commit/a0dfd94df13f0ee5097aeca92c436ade2daccdee).

I will release a new version (v1.1) with these updates.
