### 2023-01-05 #4915 @mgm8
> @mgm8 - could you please update your status and inform us on how is your corrections and edits going? Do we have any progress? Thank you in advance.

@jbytecode Sorry for the lateness. I finishing up the corrections in the implementation of the library and a new release will be published in the next few days. I'm also making improvements in the documentation.
