### 2023-01-16 mgm8/pyngham#20 @mgm8
@0xCoto [v1.1 released](https://github.com/mgm8/pyngham/releases/tag/v1.1)
