### 2023-01-17 e829705
paper: General improvements according to JOSS reviewers
