### 2023-01-17 #4915 @mgm8
> @mgm8 - A few additional suggestions for the manuscript
>
> * Line 10: Remove "Nowadays," starting the sentence with "One"
> * Line 17-28: SpaceLab is mentioned without introduction on line 17, but then properly introduced in the subsequent paragraph. Consider moving the proper introduction to the first paragraph in the Statement of Need.
> * Line 70: It is not entirely clear what a "node" is; I presume you mean a device in a network and therefore suggest using "device" or "transmitting node" or something to that effect instead.
>
> This submission is coming along nicely :)

@pritchardn done in [e829705](https://github.com/mgm8/pyngham/commit/e82970575083f80a6c20e5cf82265eb2ff325aac)! Thank you for the suggestions.
