### 2023-01-17 #4915 @0xCoto
@mgm8 Only thing remaining on my checklist is
>  [ ] Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Could you please document each function/method available in `pyngham`? Something like [this](https://virgo.readthedocs.io/en/latest/reference.html) for example, so users are aware of the available functions, how to call them, the parameter types they expect etc.
