+++ pyngham/extension.py
@@ -28,6 +28,17 @@ from pyngham.pyngham import _PYNGHAM_PL_SIZES
 class ExtPktType(Enum):
     """
     Extension packets types.
+
+    * **DATA**: Generic data packet
+    * **ID**: ID packet
+    * **STAT**: Status packet
+    * **SIMPLEDIGI**: Simple digi packet
+    * **POS**: Position data packet
+    * **TOH**: Time info packet
+    * **DEST**: Destination/receiver callsign
+    * **CMD_REQ**: Command request packet
+    * **CMD_REPLY**: Command reply packet
+    * **REQUEST**: Request packet
     """
     DATA          = 0       # Generic data packet
     ID            = 1       # ID packet
@@ -60,8 +71,10 @@ class PyNGHamExtension:
         Gets the number of extension packets in a NGHam packet.
 
         :param d: is the packet payload with extension packet(s).
+        :type d: list[int]
 
-        :return The detected number of extension packets.
+        :return: The detected number of extension packets.
+        :rtype: int
         """
         # Go through all sub packets
         start = 0
@@ -82,10 +95,15 @@ class PyNGHamExtension:
         Appends a new extension packet to a NGHam payload.
 
         :param pl: is the packet payload to append an extension packet.
-        :param typ: is the type of extension packet.
-        :param data: is the content of the extension packet.
+        :type pl: list[int]
 
-        :return The given payload with the extension packet.
+        :param typ: is the type of extension packet.
+        :type typ: int
+
+        :param data: is the content of the extension packet.
+        :type data: list[int]
+
+        :return: The given payload with the extension packet.
         """
         if len(pl) + 2 + len(data) > _PYNGHAM_PL_SIZES[-1]:
             return pl
@@ -101,10 +119,16 @@ class PyNGHamExtension:
         Adds a TOH extension packet to a NGHam payload.
 
         :param pl is the packet payload to append a TOH packet:
-        :param toh_us: time of hour in microseconds.
-        :param toh_val: validity.
+        :type pl: list[int]
 
-        :return The given payload with the new TOH extension packet.
+        :param toh_us: time of hour in microseconds.
+        :type toh_us: int
+
+        :param toh_val: validity.
+        :type toh_val: int
+
+        :return: The given payload with the new TOH extension packet.
+        :rtype: list[int]
         """
         data = list()
 
@@ -121,20 +145,46 @@ class PyNGHamExtension:
         Adds an statistic extension packet to a NGHam payload.
 
         :param pl: is the packet payload to append an statistic packet.
-        :param hw_ver: hardware version ID (10b for company, 6b for product).
-        :param serial: Serial number.
-        :param sw_ver: software version ID (4b major, 4b minor, 8b build).
-        :param uptime_s: time in whole seconds since startup.
-        :param voltage: input voltage in decivolts (0-25.5)
-        :param temp: system temperature in degrees Celsius (-128 to 127).
-        :param signal: received signal strength in dBm - 200, -200 to 54 (0xFF=N/A).
-        :param noise: noise floor, same as above.
-        :param cntr_rx_ok: packets successfully received.
-        :param cntr_rx_fix: packets with corrected errors.
-        :param cntr_rx_err: packets with uncorrectable errors.
-        :param cntr_tx: packets sent.
+        :type pl: list[int]
 
-        :return The given payload with the new statistic extension packet.
+        :param hw_ver: hardware version ID (10b for company, 6b for product).
+        :type hw_ver: int
+
+        :param serial: Serial number.
+        :type serial: int
+
+        :param sw_ver: software version ID (4b major, 4b minor, 8b build).
[…]
