### 2023-01-19 2efe918
Fixing the workflow script of the Sphinx documentation build
