+++ .github/workflows/sphinx-build.yml
@@ -14,6 +14,8 @@ jobs:
     steps:
     - uses: actions/checkout@v1
     # Standard drop-in approach that should work for most people.
+    - name: Install PyNGHam
+        run: sudo python setup.py install
     - uses: ammaraskar/sphinx-action@master
       with:
         docs-folder: "docs/"
