### 2023-01-19 9edb1bf
Adding a configuration file fir the readthedocs
