### 2023-01-19 77a9dc7
Fixing the configuration file of readthedocs
