+++ .readthedocs.yaml
@@ -28,3 +28,4 @@ python:
    install:
      - method: pip
        path: .
+     - requirements: requirements.txt
