### 2023-01-19 ef32ffe
docs: Updating the requirements.txt file
