+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 Sphinx
 sphinx-rtd-theme
+crc=3.0.0
