### 2023-01-19 7137301
docs: Updating the requirements.txt file
