+++ docs/requirements.txt
@@ -1,3 +1,3 @@
-Sphinx
-sphinx-rtd-theme
-crc=3.0.0
+Sphinx==6.1.3
+sphinx-rtd-theme==1.1.1
+crc==3.0.0
