+++ setup.py
@@ -3,7 +3,7 @@
 #
 # setup.py
 # 
-# Copyright (C) 2022, Gabriel Mariano Marcelino - PU5GMA <gabriel.mm8@gmail.com>
+# Copyright (C) 2023, Gabriel Mariano Marcelino - PU5GMA <gabriel.mm8@gmail.com>
 # 
 # This file is part of PyNGHam library.
 # 
@@ -63,5 +63,5 @@ setuptools.setup(
         ],
     download_url                    = "https://github.com/mgm8/pyngham/releases",
     packages                        = setuptools.find_packages(),
-    install_requires                = ['crc'],
+    install_requires                = ['crc==3.0.0'],
 )
