+++ requirements.txt
@@ -1,1 +1,1 @@
-crc=3.0.0
+crc==3.0.0
