### 2023-01-19 #4915 @mgm8
> @mgm8 Only thing remaining on my checklist is
>
> > [ ] Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>
> Could you please document each function/method available in `pyngham`? Something like [this](https://virgo.readthedocs.io/en/latest/reference.html) for example, so users are aware of the available functions, how to call them, the parameter types they expect etc. You've already got this as docstrings, but it would be useful to also have them available on the readthedocs page.

@0xCoto Done! Here is the reference page automatically generated from the docstrings: [https://mgm8.github.io/pyngham/reference.html](https://mgm8.github.io/pyngham/reference.html)
