+++ paper/paper.bib
@@ -105,7 +105,7 @@
 
 @misc{gr-nuts,
     author      = {Løfaldli, André},
-    title       = {gr-nuts - {GNURadio} module for the {NUTS} project groundstation},
+    title       = {{gr-nuts} - {GNU Radio} module for the {NUTS} project groundstation},
     year        = {2022},
     url         = {https://github.com/lofaldli/gr-nuts}
 }
