### 2023-01-20 #4915 @mgm8
> Dear @mgm8
>
> I've just sent a pull request [here](https://github.com/mgm8/pyngham/pull/25).
>
> Please review the changes and apply them if you are agreed with them. Please have a full read of the paper and make your corrections. Please ping me after you've done.
>
> Thank you in advance.

@jbytecode Thank you for your changes. I already applied them and also reviewed the text. I don't have any more changes to make.
