### 2023-01-20 2b764cd
Updating the list of the documentation dependencies in the main README file
