### 2023-01-20 #4915 @mgm8
@jbytecode

Done!

* Tagged release ID: **v1.1.1**
* Zenodo archive DOI: **10.5281/zenodo.7555428**
* Zenodo archive URL: [https://zenodo.org/record/7555428](https://zenodo.org/record/7555428)
