### 2023-01-21 #4915 @danielskatz
I've suggested some minor changes to the paper in https://github.com/mgm8/pyngham/pull/28 - please merge this, or let e know what you disagree with, then we can continue.
