### 2023-01-21 #4915 @mgm8
> I've suggested some minor changes to the paper in [mgm8/pyngham#28](https://github.com/mgm8/pyngham/pull/28) - please merge this, or let e know what you disagree with, then we can continue.

@danielskatz Done! I accepted the pull request!
